+++ test/support/my_app/lib/my_app_web/router.ex
@@ -59,6 +59,19 @@ defmodule MyAppWeb.Router do
     get "/scroll_props_lazy", PageController, :scroll_props_lazy
     get "/scroll_props_camelized", PageController, :scroll_props_camelized
     get "/scroll_props_with_custom_metadata", PageController, :scroll_props_with_custom_metadata
+    get "/scroll_props_scrivener", PageController, :scroll_props_scrivener
+    get "/scroll_props_flop", PageController, :scroll_props_flop
+    get "/scroll_props_with_transform", PageController, :scroll_props_with_transform
+    get "/scroll_props_with_meta", PageController, :scroll_props_with_meta
+    get "/scroll_props_invalid_meta", PageController, :scroll_props_invalid_meta
+    get "/scroll_props_legacy_transform", PageController, :scroll_props_legacy_transform
+    get "/scroll_props_camelized_wrapper", PageController, :scroll_props_camelized_wrapper
+
+    get "/scroll_props_flop_cursor_custom_metadata",
+        PageController,
+        :scroll_props_flop_cursor_custom_metadata
+
+    get "/scroll_props_invalid_tuple", PageController, :scroll_props_invalid_tuple
     get "/shared_props_via_assign", PageController, :shared_props_via_assign
     get "/shared_props_via_inline", PageController, :shared_props_via_inline
     get "/shared_props_with_merge", PageController, :shared_props_with_merge